+++ framework/model-test/src/main/java/is/codion/framework/model/test/AbstractEntityModelTest.java
@@ -90,7 +90,7 @@ public abstract class AbstractEntityModelTest<Model extends DefaultEntityModel<M
 		Entity operations = deptTableModel.selectionModel().selectedItem().get();
 		assertEquals(80, operations.get(Department.ID));
 
-		deptTableModel.includeCondition().set(item ->
+		deptTableModel.visiblePredicate().set(item ->
 						!Objects.equals(80, item.get(Department.ID)));
 
 		deptEditModel.entity().set(operations);