+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityTableModel.java
@@ -370,8 +370,8 @@ public class SwingEntityTableModel implements EntityTableModel<SwingEntityEditMo
 	}
 
 	@Override
-	public final Value<Predicate<Entity>> includeCondition() {
-		return tableModel.includeCondition();
+	public final Value<Predicate<Entity>> visiblePredicate() {
+		return tableModel.visiblePredicate();
 	}
 
 	@Override