+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/EntityComboBoxModelTest.java
@@ -210,7 +210,7 @@ public final class EntityComboBoxModelTest {
 		assertEquals(clark, comboBoxModel.selectedValue());
 		comboBoxModel.setSelectedItem(null);
 		assertNull(comboBoxModel.selectedValue());
-		comboBoxModel.includeCondition().set(entity -> false);
+		comboBoxModel.visiblePredicate().set(entity -> false);
 		comboBoxModel.select(clark.primaryKey());
 		assertEquals(clark, comboBoxModel.selectedValue());
 		Entity.Key nobodyPK = ENTITIES.primaryKey(Employee.TYPE, -1);